refactor: move ClimateNormalsPlugin into era5_land.py as ERA5LandNormalsPlugin
The normals plugin is ERA5-Land specific — it reads exclusively from the
Earth Data Hub ERA5-Land daily Zarr store and duplicated _edh_open_zarr,
_EDH_DAILY_URL, and _EDH_API_KEY_ENV already defined there. Merge it in:

- Rename to ERA5LandNormalsPlugin and add to era5_land.py
- Move dataset templates from climate_normals.yaml into era5_land.yaml
- Delete climate_normals.py and climate_normals.yaml
- Update test imports accordingly (_edh_open_daily mock → _edh_open_zarr)
