--- open_climate_service/plugins/datasets/climate_normals.yaml
@@ -1,64 +1,0 @@
-# Day-of-year climate normals (WMO 1991–2020), computed by ClimateNormalsPlugin and
-# stored as an ordinary daily dataset on a nominal leap year (2000-01-01 .. 2000-12-31).
-#
-# The built-ins below read the 30-year reference period directly from the Earth Data
-# Hub, so no prior 30-year ingestion is needed. To compute normals from an already
-# ingested managed GeoZarr store instead, use the dataset-source variant shown in the
-# commented example at the end.
-
-- id: era5land_temperature_normal
-  name: 2m temperature normal (ERA5-Land, 1991–2020)
-  short_name: Temperature normal
-  variable: t2m
-  period_type: climatology
-  sync:
-    kind: static
-  extents:
-    spatial:
-      bbox: [-180, -90, 180, 90]
-  ingestion:
-    plugin: open_climate_service.plugins.datasets.climate_normals.ClimateNormalsPlugin
-    params:
-      edh_variable: t2m
-      variable: t2m
-      period: [1991, 2020]
-      smoothing_window: 31
-      unit_transform: kelvin_to_celsius
-  units: degC
-  resolution: 9 km x 9 km
-  source: ERA5-Land Reanalysis (Earth Data Hub)
-  source_url: https://earthdatahub.destine.eu/collections/era5/datasets/era5-land-daily
-  display:
-    colormap: rdbu_r
-    range: [-30.0, 30.0]
-
-- id: era5land_precipitation_normal
-  name: Total precipitation normal (ERA5-Land, 1991–2020)
-  short_name: Precipitation normal
-  variable: tp
-  period_type: climatology
-  sync:
-    kind: static
-  extents:
-    spatial:
-      bbox: [-180, -90, 180, 90]
-  ingestion:
-    plugin: open_climate_service.plugins.datasets.climate_normals.ClimateNormalsPlugin
-    params:
-      edh_variable: tp
-      variable: tp
-      period: [1991, 2020]
-      smoothing_window: 31
-      unit_transform: metres_to_mm
-  units: mm
-  resolution: 9 km x 9 km
-  source: ERA5-Land Reanalysis (Earth Data Hub)
-  source_url: https://earthdatahub.destine.eu/collections/era5/datasets/era5-land-daily
-  display:
-    colormap: blues
-    range: [0.0, 20.0]
-
-# To compute normals from an already-ingested managed GeoZarr store (one that already
-# covers the reference period), use the openEO route instead of this plugin: a process
-# graph that loads the collection, computes the climatology, and publishes via
-# save_result. This plugin only covers the Earth Data Hub "no 30-year ingestion" path.
